Fixed GRV-1221: code sniffer violations
