+++ src/Graviton/SecurityBundle/Tests/Authentication/Strategies/AbstractHttpStrategyTest.php
@@ -63,40 +63,4 @@ class AbstractHttpStrategyTest extends GravitonSecurityBundleTestCase
             'field has invalid content (whitespaces)' => array(true, "\n\t\r\n \s"),
         );
     }
-
-}
-
-/**
- * Class AbstractHttpStrategyProxy
- *
- * @category GravitonSecurityBundle
- * @package  Graviton
- * @author   Bastian Feder <bastian.feder@swisscom.com>
- * @license  http://opensource.org/licenses/gpl-license.php GNU Public License
- * @link     http://swisscom.ch
- */
-class AbstractHttpStrategyProxy extends AbstractHttpStrategy
-{
-    /**
-     * {@inheritdoc}
-     */
-    public function extractFieldInfo($header, $fieldname)
-    {
-        return parent::extractFieldInfo($header, $fieldname);
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function validateField($header, $fieldName)
-    {
-        parent::validateField($header, $fieldName);
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function apply(Request $request)
-    {
-    }
 }
